+++ inc/hub/pages/welcome/class-welcome.php
@@ -223,7 +223,10 @@ class Welcome {
         }
 
         // Clear the old dev email option
-        delete_option( 'ddtt_dev_email' );
+        $last_viewed_version = get_option( 'ddtt_last_viewed_version', '0.0.0' );
+        if ( version_compare( $last_viewed_version, '3.0.0', '<' ) ) {
+            delete_option( 'ddtt_dev_email' );
+        }
 
         // Disable the what's new notice since we just set up
         update_option( 'ddtt_last_viewed_version', Bootstrap::version() );
